refactor: svg 경로 변경
